fix(website): fixed build and deployment for website and docs
