+++ apps/website/adapters/netlify-edge/vite.config.ts
@@ -1,7 +1,7 @@
 import { netlifyEdgeAdapter } from '@builder.io/qwik-city/adapters/netlify-edge/vite';
 import { extendConfig } from '@builder.io/qwik-city/vite';
 import { qwikVite } from '@builder.io/qwik/optimizer';
-import { UserConfig, Plugin } from 'vite';
+import { UserConfig, Plugin, UserConfigExport } from 'vite';
 import { join } from 'path';
 import baseConfig from '../../vite.config';
 
@@ -9,10 +9,10 @@ const modified: UserConfig = {
   ...baseConfig,
   // vite does not override plugins in it's "mergeConfig" util
   plugins: (baseConfig as UserConfig).plugins?.filter(
-    (p) => (p as Plugin)?.name !== 'vite-plugin-qwik',
+    (p) => (p as Plugin)?.name !== 'vite-plugin-qwik'
   ),
 };
-export default extendConfig(modified, () => {
+export default extendConfig(modified as UserConfigExport, () => {
   const outDir = 'dist/apps/website/.netlify/edge-functions/entry.netlify-edge';
   return {
     build: {
