+++ apps/website/src/speak-config.ts
@@ -1,12 +1,4 @@
-import { server$ } from '@builder.io/qwik-city';
-import {
-  LoadTranslationFn,
-  SpeakConfig,
-  TranslationFn,
-  SpeakLocale,
-  SpeakState,
-} from 'qwik-speak';
-
+import { SpeakConfig, SpeakLocale, SpeakState } from 'qwik-speak';
 export const LOCALES: Record<string, SpeakLocale> = {
   'en-US': { lang: 'en-US', currency: 'USD', timeZone: 'America/Los_Angeles' },
   'pt-BR': { lang: 'pt-BR', currency: 'BRL', timeZone: 'America/Sao_Paulo' },
@@ -33,20 +25,6 @@ export const config: SpeakConfig = {
     'subscribe',
     'showcase-page',
   ],
-};
-
-const translationData = import.meta.glob('/src/i18n/**/*.json', {
-  as: 'raw',
-  eager: true,
-});
-
-const loadTranslation$: LoadTranslationFn = server$(
-  (lang: string, asset: string) =>
-    JSON.parse(translationData[`/src/i18n/${lang}/${asset}.json`]),
-);
-
-export const translationFn: TranslationFn = {
-  loadTranslation$: loadTranslation$,
 };
 
 export const localizedUrl = (url: string, speakState: SpeakState) => {
