+++ package.json
@@ -57,6 +57,7 @@
     "react-dom": "18.2.0",
     "react-router-dom": "6.22.0",
     "regenerator-runtime": "0.14.1",
+    "sharp": "^0.33.2",
     "tapable": "2.2.1",
     "tsup": "7.2.0",
     "typedoc": "0.25.8",
@@ -199,7 +200,7 @@
     "prettier": "3.0.3",
     "prettier-eslint": "16.2.0",
     "qwik-nx": "1.1.1",
-    "qwik-speak": "0.19.0",
+    "qwik-speak": "0.12.2",
     "react-refresh": "0.14.0",
     "rimraf": "^3.0.2",
     "rollup-plugin-copy": "3.5.0",
@@ -215,7 +216,7 @@
     "url-loader": "4.1.1",
     "verdaccio": "5.29.0",
     "vinyl-fs": "4.0.0",
-    "vite": "4.4.11",
+    "vite": "5.0.10",
     "vite-tsconfig-paths": "4.2.3",
     "vitest": "1.2.2",
     "vitest-fetch-mock": "^0.2.2",
